Revert "fix connected docs"
This reverts commit d348c38954247b9bdf1bf6cf9aded8e177582236.

+++ DigitNow.Microservice.DocumentManagement/DigitNow.Domain.DocumentManagement/Business/OutgoingDocuments/Commands/Create/CreateOutgoingDocumentCommand.cs
@@ -13,6 +13,6 @@ public class CreateOutgoingDocumentCommand : ICommand<ResultObject>
     public int RecipientId { get; set; }
     public int DocumentTypeId { get; set; }
     public string? DocumentTypeDetail { get; set; }
-    public List<long>? ConnectedDocumentIds { get; set; }
+    public List<int>? ConnectedDocumentIds { get; set; }
     public List<long> UploadedFileIds { get; set; }
 }
